fix: 修复qdrant向量存储功能，新增稀疏向量存储

+++ internal/helper/config.go
@@ -8,7 +8,7 @@ import (
 
 // GetLabelThreshold 获取标签过滤阈值
 // 从配置文件读取阈值，如果未配置则使用默认值70
-func GetLabelThreshold(ctx context.Context) int {
-	threshold := g.Cfg().MustGet(ctx, "llm.label_threshold", 3).Int()
+func GetLabelThreshold(ctx context.Context) float32 {
+	threshold := g.Cfg().MustGet(ctx, "llm.label_threshold", 3).Float32()
 	return threshold
 }
